rename api to inference
