+++ tests/inference/sec10.py
@@ -9,7 +9,7 @@ from torch.utils.data import Dataset, DataLoader
 
 import zae_engine
 from zae_engine.data_pipeline.collate import Collate_seq
-from zae_engine.models.benchmark import sec10_classification
+from zae_engine.models import sec10_classification
 
 
 def core(x: np.ndarray, batch_size: int):
